UFP-584 implement group cascade delete

+++ app/src/main/java/io/unityfoundation/dds/permissions/manager/model/group/GroupService.java
@@ -9,6 +9,7 @@ import io.micronaut.http.HttpStatus;
 import io.micronaut.http.MutableHttpResponse;
 import io.micronaut.security.authentication.AuthenticationException;
 import io.unityfoundation.dds.permissions.manager.model.application.Application;
+import io.unityfoundation.dds.permissions.manager.model.applicationpermission.ApplicationPermissionRepository;
 import io.unityfoundation.dds.permissions.manager.model.groupuser.GroupUserService;
 import io.unityfoundation.dds.permissions.manager.model.topic.Topic;
 import io.unityfoundation.dds.permissions.manager.model.user.User;
@@ -24,13 +25,15 @@ import java.util.stream.Collectors;
 public class GroupService {
 
     private final GroupRepository groupRepository;
+    private final ApplicationPermissionRepository applicationPermissionRepository;
     private final SecurityUtil securityUtil;
     private final GroupUserService groupUserService;
 
 
-    public GroupService(GroupRepository groupRepository, SecurityUtil securityUtil,
+    public GroupService(GroupRepository groupRepository, ApplicationPermissionRepository applicationPermissionRepository, SecurityUtil securityUtil,
                         GroupUserService groupUserService) {
         this.groupRepository = groupRepository;
+        this.applicationPermissionRepository = applicationPermissionRepository;
         this.securityUtil = securityUtil;
         this.groupUserService = groupUserService;
     }
@@ -92,11 +95,19 @@ public class GroupService {
         }
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws Exception {
         if (!securityUtil.isCurrentUserAdmin()) {
             throw new AuthenticationException("Not authorized");
         }
+        Optional<Group> groupOptional = groupRepository.findById(id);
+        if (groupOptional.isEmpty()) {
+            throw new Exception("Group not found");
+        }
 
+        Group group = groupOptional.get();
+        groupUserService.removeByGroup(group);
+        applicationPermissionRepository.deleteByPermissionsApplicationIdIn(group.getApplications().stream().map(Application::getId).collect(Collectors.toList()));
+        applicationPermissionRepository.deleteByPermissionsTopicIdIn(group.getTopics().stream().map(Topic::getId).collect(Collectors.toList()));
         groupRepository.deleteById(id);
     }
